全面替换浮点型为金融数据类型

+++ src/klcommon/models.rs
@@ -1,4 +1,5 @@
 use serde::{Deserialize, Serialize};
+use rust_decimal::Decimal;
 use std::collections::HashMap;
 use tracing::instrument;
 
@@ -56,8 +57,8 @@ pub struct BinanceRawAggTrade {
 #[derive(Debug, Clone)]
 pub struct AppAggTrade {
     pub symbol: String,
-    pub price: f64,
-    pub quantity: f64,
+    pub price: Decimal,
+    pub quantity: Decimal,
     pub timestamp_ms: i64,
     pub is_buyer_maker: bool,
 }
